requirements.txt
fastapi
uvicorn
llama-index
llama-cpp-python
faiss-cpu
transformers
sentence-transformers
beautifulsoup4
requests
pydantic
torch